refactor: 성별 타입 중복 선언 제거 - NOTE: 현재 api 추가 없이 mock api 로 형태를 구성해 생긴 현상으로 실 기능 추가 시 수정가능
