+++ src/constants/userInfo.ts
@@ -1,3 +1,4 @@
+/** 성별 옵션 목록 */
 export const GENDER_LIST = [
   { id: "female", label: "여성" },
   { id: "male", label: "남성" },
@@ -6,6 +7,6 @@ export const GENDER_LIST = [
 
 export type GenderType = (typeof GENDER_LIST)[number]["id"];
 
-export function getGenderLabel(gender: GenderType | null): string | undefined {
-  return GENDER_LIST.find((g) => g.id === gender)?.label ?? undefined;
-}
+/** 성별 라벨 조회 */
+export const getGenderLabel = (id: string | null): string | undefined =>
+  GENDER_LIST.find((g) => g.id === id)?.label;
