+++ src/pages/main/profile/ProfileEditPage.tsx
@@ -7,8 +7,7 @@ import { authKeys } from "@/api/keyFactories";
 import { ROUTES } from "@/constants/routes";
 import { PROFILE_EDIT_TEXT } from "@/constants/texts/main/profile";
 import type { BaseResponse } from "@/api/types";
-import type { GenderType } from "@/types/auth/gender";
-import { getGenderLabel } from "@/types/auth/gender";
+import { getGenderLabel, type GenderType } from "@/constants/userInfo";
 
 import { PageTitle } from "@/components/auth/common/PageTitle";
 import { CommonInput } from "@/components/auth/common/CommonInput";
@@ -72,18 +71,7 @@ const ProfileEditPage = () => {
   useEffect(() => {
     if (userData) {
       setNickname(userData.nickName || "");
-
-      // 성별 변환: API에서 "M"/"F" 형태로 올 수 있음
-      if (userData.gender) {
-        const genderMap: Record<string, GenderType> = {
-          M: "male",
-          F: "female",
-          male: "male",
-          female: "female",
-          other: "other",
-        };
-        setGender(genderMap[userData.gender] || null);
-      }
+      setGender((userData.gender as GenderType) || null);
 
       // 생년월일 파싱: "YYYYMMDD" 형식
       if (userData.birth && userData.birth.length === 8) {
@@ -127,13 +115,6 @@ const ProfileEditPage = () => {
 
   // 프로필 수정 제출
   const handleSubmitProfile = () => {
-    // 성별 변환: API 규격에 맞춤
-    const genderMap: Record<GenderType, string> = {
-      male: "M",
-      female: "F",
-      other: "O",
-    };
-
     const birth =
       userBirthYear && userBirthMonth && userBirthDay
         ? `${userBirthYear}${userBirthMonth}${userBirthDay}`
@@ -141,7 +122,7 @@ const ProfileEditPage = () => {
 
     updateMutation.mutate({
       nickName: nickname.trim(),
-      gender: gender ? genderMap[gender] : undefined,
+      gender: gender ?? undefined,
       birth,
     });
   };
